comment pre-commit code line
