+++ backend/organization/tests/test_project_collection_views.py
@@ -1,28 +1,16 @@
-import io
 import unittest
-from base64 import b64encode
 from unittest.mock import patch
 
-from django.contrib.auth.models import User
 from django.contrib.gis.geos import MultiPolygon, Point, Polygon
 from django.test import TestCase, tag
 from django.urls import reverse
-from PIL import Image
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from climateconnect_api.models import Language, Role
+from climateconnect_api.models import Language
 from hubs.models.hub import Hub
-from location.models import Location, LocationTranslation
-from organization.models import (
-    Organization,
-    Project,
-    ProjectMember,
-    ProjectParents,
-    ProjectSectorMapping,
-    ProjectStatus,
-    Sector,
-)
+from location.models import Location
+from organization.models import Project, ProjectSectorMapping, ProjectStatus, Sector
 
 # set this at lowest to 4
 INITIAL_PROJECT_COUNT = 4
@@ -68,22 +56,17 @@ class TestProjectsListView(APITestCase):
 
     @tag("projects")
     def test_get_projects_lists_content(self):
-        # arrange
-
-        # act
         response = self.client.get(self.url)
         results = response.json().get("results", None)
 
         for i in range(1, INITIAL_PROJECT_COUNT + 1):
             self.assertContains(response, f"Test Project {i}")
 
-        # assert
         self.assertIsNotNone(results)
         self.assertEqual(len(results), INITIAL_PROJECT_COUNT)
 
     @tag("sectors", "projects")
     def test_get_projects_lists_content_filtered_by_a_single_sector(self):
-        # arranging
         sectors = [
             Sector.objects.create(
                 name=f"Test Sector Name {i}",
@@ -95,13 +78,10 @@ class TestProjectsListView(APITestCase):
 
         projects = self.projects
 
-        # testing size >1, =1 and =0 for the result length
         ProjectSectorMapping.objects.create(sector=sectors[0], project=projects[0])
         ProjectSectorMapping.objects.create(sector=sectors[0], project=projects[1])
         ProjectSectorMapping.objects.create(sector=sectors[1], project=projects[2])
-        # sector 3 will not be connected to any project
 
-        # act
         response_1 = self.client.get(self.url + "?sectors=" + sectors[0].name)
         response_2 = self.client.get(self.url + "?sectors=" + sectors[1].name)
         response_3 = self.client.get(self.url + "?sectors=" + sectors[2].name)
@@ -110,22 +90,20 @@ class TestProjectsListView(APITestCase):
         results_2 = response_2.json().get("results", None)
         results_3 = response_3.json().get("results", None)
 
-        # assert
         self.assertIsNotNone(results_1)
         self.assertEqual(len(results_1), 2)
-        self.assertContains(response_1, f"Test Project {1}")
-        self.assertContains(response_1, f"Test Project {2}")
+        self.assertContains(response_1, "Test Project 1")
+        self.assertContains(response_1, "Test Project 2")
 
         self.assertIsNotNone(results_2)
         self.assertEqual(len(results_2), 1)
-        self.assertContains(response_2, f"Test Project {3}")
+        self.assertContains(response_2, "Test Project 3")
 
         self.assertIsNotNone(results_3)
         self.assertEqual(len(results_3), 0)
 
     @tag("sectors", "projects")
     def test_get_projects_lists_content_filtered_by_two_sectors(self):
-        # arranging
         sectors = [
             Sector.objects.create(
                 name=f"Test Sector Name {i}",
@@ -159,18 +137,14 @@ class TestProjectsListView(APITestCase):
             },
             {
                 "url": "sectors={},{}".format(sectors[1].name, sectors[2].name),
-                "expected": [
-                    projects[2].url_slug,
-                ],
+                "expected": [projects[2].url_slug],
             },
         ]
 
-        # act
         responses = [
             self.client.get(self.url + "?" + testcase["url"]) for testcase in testcases
         ]
 
-        # assert
         for i, response in enumerate(responses):
             results = response.json().get("results", None)
 
@@ -182,7 +156,6 @@ class TestProjectsListView(APITestCase):
 
     @tag("sectors", "projects")
     def test_get_projects_lists_content_filtered_by_sector_hub(self):
-        # arrange
         sectors = [
             Sector.objects.create(
                 name=f"Test Sector Name {i}",
@@ -193,13 +166,10 @@ class TestProjectsListView(APITestCase):
         ]
 
         projects = self.projects
-        ## mapping projects to sectors
-        ## testing number of projects per sector: 2, 1 and 0 (testing a loop)
         ProjectSectorMapping.objects.create(sector=sectors[0], project=projects[0])
         ProjectSectorMapping.objects.create(sector=sectors[0], project=projects[1])
         ProjectSectorMapping.objects.create(sector=sectors[1], project=projects[2])
 
-        # linking sectors to sector hubs, to query projects by sector hubs
         self.sector_hubs = [
             Hub.objects.create(
                 name="Test sector Hub {}".format(i),
@@ -218,19 +188,13 @@ class TestProjectsListView(APITestCase):
         )
 
         testcases = [
-            # test query by a single sector hub
             {
                 "url": "hub={}".format(self.sector_hubs[0].url_slug),
-                "expected": [
-                    projects[0].url_slug,
-                    projects[1].url_slug,
-                ],
+                "expected": [projects[0].url_slug, projects[1].url_slug],
             },
             {
                 "url": "hub={}".format(self.sector_hubs[1].url_slug),
-                "expected": [
-                    projects[2].url_slug,
-                ],
+                "expected": [projects[2].url_slug],
             },
             {
                 "url": "hub={}".format(self.sector_hubs[2].url_slug),
@@ -242,12 +206,10 @@ class TestProjectsListView(APITestCase):
             },
         ]
 
-        # act
         responses = [
             self.client.get(self.url + "?" + testcases[i]["url"]) for i in range(3)
         ]
 
-        # assert
         for i, response in enumerate(responses):
             results = response.json().get("results", None)
 
@@ -256,11 +218,9 @@ class TestProjectsListView(APITestCase):
 
             for project_slug in testcases[i]["expected"]:
                 self.assertContains(response, project_slug)
-        pass
 
     @tag("sectors", "projects")
     def test_get_project_sector_ordering(self):
-        # arrange
         sectors = [
             Sector.objects.create(
                 name=f"Test Sector Name {i}",
@@ -271,30 +231,18 @@ class TestProjectsListView(APITestCase):
         ]
 
         projects = self.projects
-        ## mapping projects to sectors
-        ## Setup orderings for sector 0
-        ordering_0 = [
-            (0, 1),
-            (1, 2),
-            (2, 3),
-        ]
+        ordering_0 = [(0, 1), (1, 2), (2, 3)]
         for x, y in ordering_0:
             ProjectSectorMapping.objects.create(
                 project=projects[0], sector=sectors[x], order=y
             )
 
-        ## Setup orderings for sector 1
-        ordering_1 = [
-            (2, 1),
-            (1, 2),
-            (0, 3),
-        ]
+        ordering_1 = [(2, 1), (1, 2), (0, 3)]
         for x, y in ordering_1:
             ProjectSectorMapping.objects.create(
                 project=projects[1], sector=sectors[x], order=y
             )
 
-        # act
         response = self.client.get(self.url + "?sectors=" + sectors[0].name)
         result = response.json().get("results", None)
         self.assertIsNotNone(result)
@@ -308,7 +256,6 @@ class TestProjectsListView(APITestCase):
             if _project["url_slug"] == self.projects[1].url_slug:
                 project_1 = _project
 
-        # assert
         self.assertIsNotNone(project_0)
         self.assertIsNotNone(project_1)
 
@@ -322,913 +269,12 @@ class TestProjectsListView(APITestCase):
         self.assertEqual(len(sectors_1), 3)
 
         for i, (x, y) in enumerate(ordering_0):
-            self.assertEqual(
-                sectors_0[i]["sector"]["key"],
-                sectors[x].key,
-            )
+            self.assertEqual(sectors_0[i]["sector"]["key"], sectors[x].key)
             self.assertEqual(sectors_0[i]["order"], y)
 
         for i, (x, y) in enumerate(ordering_1):
             self.assertEqual(sectors_1[i]["sector"]["key"], sectors[x].key)
             self.assertEqual(sectors_1[i]["order"], y)
-
-
-@unittest.skip("Temporarily disabled: see CI failure #57660401767")
-class TestCreateProjectsViews(APITestCase):
-    # -----------------------------------------------------
-    # setUp code for each test
-    def setUp(self):
-        self.url = reverse("organization:create-project-api")
-
-        projectStatus_active = ProjectStatus.objects.create(
-            name="active",
-            name_de_translation="aktiv",
-            has_end_date=False,
-            has_start_date=False,
-        )
-
-        self.proejcts = [
-            Project.objects.create(
-                name=f"Test Project {i}",
-                description=f"Test Project {i} Description",
-                url_slug=f"test-project-{i}",
-                is_active=True,
-                status=projectStatus_active,
-            )
-            for i in range(1, INITIAL_PROJECT_COUNT + 1)
-        ]
-
-        self.user = User.objects.create_user(
-            username="testuser",
-            password="testpassword",
-        )
-
-        Location.objects.create(
-            name="Test Location", city="Berlin", country="Germany", place_id=1
-        )
-
-        # Get the language that was created globally
-        self.default_language = Language.objects.get(language_code="en")
-
-        self.default_location_data = {
-            "place_id": 1,
-            "country": "Germany",
-            "name": "Test Location",
-            "type": "type",
-            "lon": 13.4050,
-            "lat": 52.5200,
-        }
-        # create a dummy black image
-        image = Image.new("RGB", (10, 10), "black")
-        buffered = io.BytesIO()
-        image.save(buffered, format="PNG")
-        self.image = "data:image/png;base64," + b64encode(buffered.getvalue()).decode(
-            "utf-8"
-        )
-
-        # TODO: broooooo, why -__-
-        # we need to change the api asap to just accept the 'project_type'
-        # and not the whole object ...
-        self.default_project_type_data = {
-            "name": "Project",
-            "original_name": "Project",
-            "help_text": "Not an Idea or Event? Click here.",
-            "icon": "",
-            "type_id": "project",
-        }
-
-        self.default_project_data = {
-            "name": "Test Project",
-            "status": ProjectStatus.objects.first().id,
-            "short_description": "Test Project Short Description",
-            "collaborators_welcome": True,
-            "team_members": [],
-            "description": "Test Project Description",
-            "url_slug": "test-project",
-            "project_tags": [],
-            "loc": self.default_location_data,
-            "image": self.image,
-            "source_language": self.default_language.language_code,
-            "translations": {},
-            "project_type": self.default_project_type_data,
-            "hubName": "null",
-        }
-
-        # creating dummy sectors
-
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector Name {i}",
-                name_de_translation=f"Test Sector Name {i} DE",
-                key=f"test_sector_{i}",
-            )
-            for i in range(3)
-        ]
-
-    def test_post_project_declines_if_not_authenticated(self):
-        # arrange
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    @tag("projects")
-    def test_post_project_without_sector(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            0,
-        )
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_one_sector_as_array(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.default_project_data["sectors"] = [self.sectors[0].key]
-
-        # act
-        self.client.login(username="testuser", password="testpassword")
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            1,
-        )
-
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_one_sector_as_string(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-        self.default_project_data["sectors"] = self.sectors[0].key
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            1,
-        )
-
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_two_sector_list(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.default_project_data["sectors"] = [
-            self.sectors[0].key,
-            self.sectors[1].key,
-        ]
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            2,
-        )
-
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_two_sector_comma_seperated(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.default_project_data[
-            "sectors"
-        ] = f"{self.sectors[0].key},{self.sectors[1].key}"
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            2,
-        )
-
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_two_sector_dublicated(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.default_project_data[
-            "sectors"
-        ] = f"{self.sectors[0].key},{self.sectors[0].key}"
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.default_project_data["url_slug"]
-            ).count(),
-            1,
-        )
-
-        pass
-
-    @tag("sectors", "projects")
-    def test_post_project_with_sectors_will_keep_the_correct_ordering(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.default_project_data["sectors"] = ",".join([s.key for s in self.sectors])
-
-        # act
-        response = self.client.post(self.url, self.default_project_data, format="json")
-
-        # assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        mappings = ProjectSectorMapping.objects.filter(
-            project__url_slug=self.default_project_data["url_slug"]
-        )
-        self.assertEqual(len(self.sectors), mappings.count())
-        n = len(self.sectors)
-        for i, mapping in enumerate(mappings):
-            expected_order = n - i
-            self.assertEqual(expected_order, mapping.order)
-        pass
-
-
-class TestProjectApi(APITestCase):
-    def setUp(self):
-        self.url_slug = "test-project"
-        self.url = reverse(
-            "organization:project-api-view", kwargs={"url_slug": "test-project"}
-        )
-        self.edit_view_url = self.url + "?" + "edit_view"
-
-        self.projectStatus_active = ProjectStatus.objects.create(
-            name="active",
-            name_de_translation="aktiv",
-            has_end_date=False,
-            has_start_date=False,
-        )
-
-        # Get the language that was created globally
-        self.default_language = Language.objects.get(language_code="en")
-
-        self.project = Project.objects.create(
-            name="Test Project",
-            description="Test Project Description",
-            url_slug=self.url_slug,
-            is_active=True,
-            status=self.projectStatus_active,
-            language=self.default_language,
-        )
-
-        self.decoy_project = Project.objects.create(
-            name="decoy",
-            description="decoy desc",
-            url_slug="decoy",
-            is_active=True,
-            status=self.projectStatus_active,
-        )
-
-        self.user = User.objects.create_user(
-            username="testuser",
-            password="testpassword",
-        )
-
-        self.role = Role.objects.create(
-            name="Admin",
-            role_type=Role.ALL_TYPE,
-        )
-        ProjectMember.objects.create(
-            user=self.user,
-            project=self.project,
-            role=self.role,
-        )
-        ProjectParents.objects.create(
-            project=self.project,
-            parent_user=self.user,
-        )
-
-    @tag("projects")
-    def test_get_project_by_url_slug(self):
-        # arrange
-
-        # act
-        response = self.client.get(self.url)
-        res = response.json()
-
-        # assert
-        self.assertIsNotNone(res)
-        self.assertContains(response, "Test Project")
-        self.assertNotContains(response, "decoy")
-
-    @tag("projects", "location")
-    def test_get_project_by_url_slug_returns_translated_location_name(self):
-        location = Location.objects.create(
-            name="Munich",
-            city="Munich",
-            country="Germany",
-        )
-        LocationTranslation.objects.create(
-            location=location,
-            language=Language.objects.get(language_code="de"),
-            name_translation="München",
-        )
-        self.project.loc = location
-        self.project.save()
-
-        response = self.client.get(self.url, HTTP_ACCEPT_LANGUAGE="de")
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.json()["location"], "München")
-
-    @tag("projects", "location")
-    def test_get_project_by_url_slug_falls_back_to_english_location_name(self):
-        location = Location.objects.create(
-            name="Cologne",
-            city="Cologne",
-            country="Germany",
-        )
-        self.project.loc = location
-        self.project.save()
-
-        response = self.client.get(self.url, HTTP_ACCEPT_LANGUAGE="de")
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.json()["location"], "Cologne")
-
-    @tag("sectors", "projects")
-    def test_get_project_by_url_slug_includes_sector(self):
-        # arrange
-        self.sector = Sector.objects.create(
-            name="Test Sector",
-            name_de_translation="Test Sector DE",
-            key="test_sector",
-        )
-        self.sector_decoy = Sector.objects.create(
-            name="Test Sector decoy",
-            name_de_translation="Test Sector DE decoy",
-            key="test_sector_decoy",
-        )
-
-        ProjectSectorMapping.objects.create(sector=self.sector, project=self.project)
-        ProjectSectorMapping.objects.create(
-            sector=self.sector_decoy, project=self.decoy_project
-        )
-
-        # act
-        response = self.client.get(self.url)
-        res = response.json()
-
-        # assert
-        self.assertIsNotNone(res)
-        self.assertContains(response, "Test Project")
-        self.assertContains(response, "Test Sector")
-        self.assertNotContains(response, "decoy")
-
-    @tag("sectors", "projects")
-    def test_get_project_edit_view_by_url_slug_includes_sector(self):
-        # arrange
-        self.sector = Sector.objects.create(
-            name="Test Sector",
-            name_de_translation="Test Sector DE",
-            key="test_sector",
-        )
-        self.sector_decoy = Sector.objects.create(
-            name="Test Sector decoy",
-            name_de_translation="Test Sector DE decoy",
-            key="test_sector_decoy",
-        )
-
-        ProjectSectorMapping.objects.create(sector=self.sector, project=self.project)
-        ProjectSectorMapping.objects.create(
-            sector=self.sector_decoy, project=self.decoy_project
-        )
-
-        # act
-        response = self.client.get(self.edit_view_url)
-        res = response.json()
-
-        # assert
-        self.assertIsNotNone(res)
-        self.assertContains(response, "Test Project")
-        self.assertContains(response, "Test Sector")
-        self.assertNotContains(response, "decoy")
-
-    @tag("sectors", "projects")
-    def test_get_project_by_url_slug_includes_sector_correctly_sorted(self):
-        # arrange
-        N = 4
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector{i}",
-                name_de_translation=f"Test Sector DE {i}",
-                key=f"test_sector_{i}",
-            )
-            for i in range(N)
-        ]
-        ordering = {self.sectors[i].key: i + 1 for i in range(N)}
-
-        for sector in self.sectors:
-            ProjectSectorMapping.objects.create(
-                sector=sector, project=self.project, order=ordering[sector.key]
-            )
-
-        # act
-        response = self.client.get(self.url)
-        res_view = response.json().get("sectors", None)
-
-        response_edit = self.client.get(self.edit_view_url)
-        res_edit = response_edit.json().get("sectors", None)
-
-        # assert
-        ## perform it for both results, as the ordering should be the same on both views
-        for res in [res_view, res_edit]:
-            self.assertIsNotNone(res)
-            for item in res:
-                sector = item.get("sector", None)
-                order = item.get("order", None)
-
-                self.assertIsNotNone(sector)
-                self.assertIsNotNone(order)
-
-                key = sector["key"]
-                order = int(order)
-                expected_order = ordering[key]
-
-                self.assertIsNotNone(expected_order)
-                self.assertEqual(order, expected_order)
-
-    @tag("sectors", "projects")
-    def test_patch_project_adding_first_sector(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.sector = Sector.objects.create(
-            name="Test Sector",
-            name_de_translation="Test Sector DE",
-            key="test_sector",
-        )
-
-        data = {
-            "sectors": [self.sector.key],
-        }
-
-        # act
-        self.client.patch(self.url, data, format="json")
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            1,
-        )
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug, sector__key=self.sector.key
-            ).count(),
-            1,
-        )
-
-    @tag("sectors", "projects")
-    def test_patch_project_adding_another_sectors(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector {i}",
-                name_de_translation=f"Test Sector DE {i}",
-                key=f"test_sector_{i}",
-            )
-            for i in range(2)
-        ]
-        ProjectSectorMapping.objects.create(
-            sector=self.sectors[0], project=self.project
-        )
-
-        data = {
-            "sectors": [self.sectors[0].key, self.sectors[1].key],
-        }
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            2,
-        )
-        for sector in self.sectors:
-            self.assertEqual(
-                ProjectSectorMapping.objects.filter(
-                    project__url_slug=self.project.url_slug, sector__key=sector.key
-                ).count(),
-                1,
-            )
-
-    @tag("sectors", "projects")
-    def test_patch_project_adding_multiple_sectors(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        ## set up another project with the same user as an admin
-        self.other_project = Project.objects.create(
-            name="Other Project",
-            description="Other Project Description",
-            url_slug="other-project",
-            is_active=True,
-            status=self.projectStatus_active,
-            language=self.default_language,
-        )
-        self.other_url = self.url.replace(
-            self.project.url_slug, self.other_project.url_slug
-        )
-        ProjectMember.objects.create(
-            user=self.user,
-            project=self.other_project,
-            role=self.role,
-        )
-
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector {i}",
-                name_de_translation=f"Test Sector DE {i}",
-                key=f"test_sector_{i}",
-            )
-            for i in range(4)
-        ]
-        ProjectSectorMapping.objects.create(
-            sector=self.sectors[0], project=self.other_project
-        )
-        ProjectSectorMapping.objects.create(
-            sector=self.sectors[3], project=self.decoy_project
-        )
-
-        data = {
-            "sectors": [
-                self.sectors[0].key,
-                self.sectors[1].key,
-                self.sectors[2].key,
-            ],
-        }
-        data_other = {
-            "sectors": [
-                self.sectors[0].key,
-                self.sectors[1].key,
-                self.sectors[3].key,
-            ],
-        }
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-        response_other = self.client.patch(self.other_url, data_other, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertContains(response_other, "successfully updated")
-
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            3,
-        )
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.other_project.url_slug
-            ).count(),
-            3,
-        )
-
-        for sector in self.sectors[:3]:
-            self.assertEqual(
-                ProjectSectorMapping.objects.filter(
-                    project__url_slug=self.project.url_slug, sector__key=sector.key
-                ).count(),
-                1,
-            )
-
-        for sector in [self.sectors[0], self.sectors[1], self.sectors[3]]:
-            self.assertEqual(
-                ProjectSectorMapping.objects.filter(
-                    project__url_slug=self.other_project.url_slug,
-                    sector__key=sector.key,
-                ).count(),
-                1,
-            )
-
-    @tag("sectors", "projects")
-    def test_patch_project_removing_sector(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.sector = Sector.objects.create(
-            name="Test Sector",
-            name_de_translation="Test Sector DE",
-            key="test_sector",
-        )
-        ProjectSectorMapping.objects.create(sector=self.sector, project=self.project)
-
-        data = {
-            "sectors": [],
-        }
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            0,
-        )
-
-    @tag("sectors", "projects")
-    def test_patch_project_ordering_of_sectors_correctly_assigned(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector {i}",
-                name_de_translation=f"Test Sector DE {i}",
-                key=f"test_sector_{i}",
-            )
-            for i in range(4)
-        ]
-        ProjectSectorMapping.objects.create(
-            sector=self.sectors[0], project=self.project, order=1
-        )
-
-        data = {
-            "sectors": [self.sectors[3].key, self.sectors[2].key, self.sectors[1].key]
-        }
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            3,
-        )
-
-        for i, sector in enumerate(self.sectors[1:]):
-            mappings = ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug, sector__key=sector.key
-            )
-            self.assertEqual(mappings.count(), 1)
-            mapping = mappings[0]
-            self.assertEqual(mapping.order, i + 1)
-
-    @tag("sectors", "projects")
-    def test_patch_project_reordering_of_sectors(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-
-        N = 4
-        self.sectors = [
-            Sector.objects.create(
-                name=f"Test Sector {i}",
-                name_de_translation=f"Test Sector DE {i}",
-                key=f"test_sector_{i}",
-            )
-            for i in range(N)
-        ]
-
-        for i, sector in enumerate(self.sectors):
-            ProjectSectorMapping.objects.create(
-                sector=sector, project=self.project, order=i
-            )
-
-        data = {
-            "sectors": [self.sectors[N - i - 1].key for i in range(N)]
-            + [s.key for s in self.sectors],
-        }
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            N,
-        )
-
-        for i, sector in enumerate(self.sectors):
-            mapping = ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug, sector__key=sector.key
-            ).first()
-            self.assertEqual(mapping.order, i + 1)
-
-    @tag("sectors", "projects")
-    def test_delete_project_sector(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-        self.sector = Sector.objects.create(
-            name="Test Sector",
-            name_de_translation="Test Sector DE",
-            key="test_sector",
-        )
-        ProjectSectorMapping.objects.create(sector=self.sector, project=self.project)
-        data = {
-            "sectors": [],
-        }
-
-        # act
-        response = self.client.delete(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully deleted")
-        self.assertEqual(
-            ProjectSectorMapping.objects.filter(
-                project__url_slug=self.project.url_slug
-            ).count(),
-            0,
-        )
-
-    # -------------------------------------------------------
-    # Tests for the is_online field
-
-    @tag("projects")
-    def test_get_project_includes_is_online_field(self):
-        # act
-        response = self.client.get(self.url)
-        res = response.json()
-
-        # assert
-        self.assertIn("is_online", res)
-
-    @tag("projects")
-    def test_get_project_is_online_defaults_to_false(self):
-        # act
-        response = self.client.get(self.url)
-        res = response.json()
-
-        # assert
-        self.assertFalse(res["is_online"])
-
-    @tag("projects")
-    def test_get_project_is_online_true_when_set(self):
-        # arrange
-        self.project.is_online = True
-        self.project.save()
-
-        # act
-        response = self.client.get(self.url)
-        res = response.json()
-
-        # assert
-        self.assertTrue(res["is_online"])
-
-    @tag("projects")
-    def test_patch_project_set_is_online_to_true(self):
-        # arrange
-        self.client.login(username="testuser", password="testpassword")
-        data = {"is_online": True}
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.project.refresh_from_db()
-        self.assertTrue(self.project.is_online)
-
-    @tag("projects")
-    def test_patch_project_set_is_online_to_false(self):
-        # arrange
-        self.project.is_online = True
-        self.project.save()
-        self.client.login(username="testuser", password="testpassword")
-        data = {"is_online": False}
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.project.refresh_from_db()
-        self.assertFalse(self.project.is_online)
-
-    @tag("projects")
-    def test_patch_project_without_is_online_does_not_change_default(self):
-        # arrange — project starts with is_online=False (default)
-        self.client.login(username="testuser", password="testpassword")
-        data = {"website": "https://example.com"}
-
-        # act
-        response = self.client.patch(self.url, data, format="json")
-
-        # assert
-        self.assertContains(response, "successfully updated")
-        self.project.refresh_from_db()
-        self.assertFalse(self.project.is_online)
-
-    @tag("projects")
-    def test_get_project_edit_view_includes_is_online_field(self):
-        # act
-        response = self.client.get(self.edit_view_url)
-        res = response.json()
-
-        # assert
-        self.assertIn("is_online", res)
-
-    @tag("projects")
-    def test_patch_project_keeps_organization_owner_when_is_personal_project_flag_is_stale(self):
-        self.client.login(username="testuser", password="testpassword")
-
-        organization = Organization.objects.create(
-            name="Owner Organization",
-            url_slug="owner-organization-for-stale-flag-test",
-            language=self.default_language,
-        )
-
-        project_parents = self.project.project_parent.get()
-        project_parents.parent_organization = organization
-        project_parents.save()
-
-        response = self.client.patch(
-            self.url,
-            {"name": "Updated Project Name", "is_personal_project": True},
-            format="json",
-        )
-
-        self.assertContains(response, "successfully updated")
-        project_parents.refresh_from_db()
-        self.assertEqual(project_parents.parent_organization, organization)
-        self.project.refresh_from_db()
-        self.assertEqual(self.project.name, "Updated Project Name")
-
-    @tag("projects")
-    def test_patch_project_rejects_invalid_parent_organization_id(self):
-        self.client.login(username="testuser", password="testpassword")
-
-        project_parents = self.project.project_parent.get()
-        project_parents.parent_organization = Organization.objects.create(
-            name="Existing Organization",
-            url_slug="existing-organization-for-invalid-id-test",
-            language=self.default_language,
-        )
-        project_parents.save()
-
-        response = self.client.patch(
-            self.url,
-            {"parent_organization": 999999},
-            format="json",
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn("parent_organization", response.json())
-        project_parents.refresh_from_db()
-        self.assertIsNotNone(project_parents.parent_organization)
 
 
 class ProjectLocationHubFilterTest(TestCase):
@@ -1239,26 +285,8 @@ class ProjectLocationHubFilterTest(TestCase):
     """
 
     def setUp(self):
-        """
-        Set up test data for each test method.
-        Creates locations, hubs, and projects for testing.
-
-        Test data structure:
-        - 1 Hub "Erlangen" with 3 locations: Erlangen, Bubenreuth, Spardorf
-        - 12 Projects total:
-          - 3 with exact hub locations (Erlangen, Bubenreuth, Spardorf)
-          - 3 with addresses in Erlangen, Bubenreuth, Spardorf
-          - 1 inactive project (is_active=False)
-          - 1 draft project (is_draft=True)
-          - 2 in Nürnberg (location + address) - negative test cases
-          - 2 in Paris (location + address) - negative test cases
-        """
         self.url = reverse("organization:list-projects")
-
-        # Create language (required for projects)
         self.language = Language.objects.get(language_code="en")
-
-        # Create project status (required for projects)
         self.project_status = ProjectStatus.objects.create(
             name="active",
             name_de_translation="aktiv",
@@ -1266,13 +294,11 @@ class ProjectLocationHubFilterTest(TestCase):
             has_start_date=False,
         )
 
-        # ===== Hub Locations (3 locations for Erlangen Hub) =====
-        # Erlangen city (general area)
         self.location_erlangen = Location.objects.create(
             name="Erlangen, Germany",
             city="Erlangen",
             country="Germany",
-            centre_point=Point(11.0050, 49.5975),  # Erlangen coordinates
+            centre_point=Point(11.0050, 49.5975),
             multi_polygon=MultiPolygon(
                 Polygon(
                     (
@@ -1285,13 +311,11 @@ class ProjectLocationHubFilterTest(TestCase):
                 )
             ),
         )
-
-        # Bubenreuth (suburb/village north of Erlangen)
         self.location_bubenreuth = Location.objects.create(
             name="Bubenreuth, Germany",
             city="Bubenreuth",
             country="Germany",
-            centre_point=Point(11.0200, 49.6300),  # Bubenreuth coordinates
+            centre_point=Point(11.0200, 49.6300),
             multi_polygon=MultiPolygon(
                 Polygon(
                     (
@@ -1304,13 +328,11 @@ class ProjectLocationHubFilterTest(TestCase):
                 )
             ),
         )
-
-        # Spardorf (suburb/village east of Erlangen)
         self.location_spardorf = Location.objects.create(
             name="Spardorf, Germany",
             city="Spardorf",
             country="Germany",
-            centre_point=Point(11.0600, 49.5900),  # Spardorf coordinates
+            centre_point=Point(11.0600, 49.5900),
             multi_polygon=MultiPolygon(
                 Polygon(
                     (
@@ -1324,7 +346,6 @@ class ProjectLocationHubFilterTest(TestCase):
             ),
         )
 
-        # ===== Hub for Erlangen with 3 locations =====
         self.hub_erlangen = Hub.objects.create(
             name="Erlangen Hub",
             url_slug="erlangen-hub",
@@ -1335,39 +356,33 @@ class ProjectLocationHubFilterTest(TestCase):
         self.hub_erlangen.location.add(self.location_bubenreuth)
         self.hub_erlangen.location.add(self.location_spardorf)
 
-        # ===== Address Locations within hub area =====
-        # Addresses within Erlangen, Bubenreuth, Spardorf
         self.location_erlangen_address = Location.objects.create(
             name="Goethestraße 1, Erlangen, Germany",
             city="Erlangen",
             country="Germany",
-            centre_point=Point(11.0020, 49.5985),  # Address in Erlangen
+            centre_point=Point(11.0020, 49.5985),
             multi_polygon=None,
         )
-
         self.location_bubenreuth_address = Location.objects.create(
             name="Bussardstraße 21, Bubenreuth, Germany",
             city="Bubenreuth",
             country="Germany",
-            centre_point=Point(11.0210, 49.6310),  # Address in Bubenreuth
+            centre_point=Point(11.0210, 49.6310),
             multi_polygon=None,
         )
-
         self.location_spardorf_address = Location.objects.create(
             name="Eisenstraße, Spardorf, Germany",
             city="Spardorf",
             country="Germany",
-            centre_point=Point(11.0610, 49.5910),  # Address in Spardorf
+            centre_point=Point(11.0610, 49.5910),
             multi_polygon=None,
         )
 
-        # ===== Negative Test Locations (Nuremberg & Paris) =====
-        # Nuremberg city location
         self.location_nuremberg = Location.objects.create(
             name="Nuremberg, Germany",
             city="Nuremberg",
             country="Germany",
-            centre_point=Point(11.0767, 49.4521),  # Nuremberg coordinates
+            centre_point=Point(11.0767, 49.4521),
             multi_polygon=MultiPolygon(
                 Polygon(
                     (
@@ -1380,22 +395,18 @@ class ProjectLocationHubFilterTest(TestCase):
                 )
             ),
         )
-
-        # Nuremberg address
         self.location_nuremberg_address = Location.objects.create(
             name="Narrenschiff, Plobenhofstraße 1-9, Nuremberg, Germany",
             city="Nuremberg",
             country="Germany",
-            centre_point=Point(11.0780, 49.4530),  # Address in Nuremberg
+            centre_point=Point(11.0780, 49.4530),
             multi_polygon=None,
         )
-
-        # Paris city location
         self.location_paris = Location.objects.create(
             name="Paris, France",
             city="Paris",
             country="France",
-            centre_point=Point(2.3522, 48.8566),  # Paris coordinates
+            centre_point=Point(2.3522, 48.8566),
             multi_polygon=MultiPolygon(
                 Polygon(
                     (
@@ -1408,18 +419,14 @@ class ProjectLocationHubFilterTest(TestCase):
                 )
             ),
         )
-
-        # Paris address
         self.location_paris_address = Location.objects.create(
             name="Hotel de Ville, 5 Rue de Lobau, Paris, France",
             city="Paris",
             country="France",
-            centre_point=Point(2.3530, 48.8570),  # Address in Paris
+            centre_point=Point(2.3530, 48.8570),
             multi_polygon=None,
         )
 
-        # ===== Create 12 Projects =====
-        # Projects with exact hub locations (3 projects)
         self.project_erlangen_location = Project.objects.create(
             name="Erlangen Location Project",
             description="Project with Erlangen location",
@@ -1431,7 +438,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_erlangen,
         )
-
         self.project_bubenreuth_location = Project.objects.create(
             name="Bubenreuth Location Project",
             description="Project with Bubenreuth location",
@@ -1443,7 +449,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_bubenreuth,
         )
-
         self.project_spardorf_location = Project.objects.create(
             name="Spardorf Location Project",
             description="Project with Spardorf location",
@@ -1455,8 +460,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_spardorf,
         )
-
-        # Projects with addresses in hub locations (3 projects)
         self.project_erlangen_address = Project.objects.create(
             name="Erlangen Address Project",
             description="Project with address in Erlangen",
@@ -1468,7 +471,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_erlangen_address,
         )
-
         self.project_bubenreuth_address = Project.objects.create(
             name="Bubenreuth Address Project",
             description="Project with address in Bubenreuth",
@@ -1480,7 +482,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_bubenreuth_address,
         )
-
         self.project_spardorf_address = Project.objects.create(
             name="Spardorf Address Project",
             description="Project with address in Spardorf",
@@ -1492,8 +493,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_spardorf_address,
         )
-
-        # Inactive project
         self.project_erlangen_inactive = Project.objects.create(
             name="Erlangen Inactive Project",
             description="Inactive project in Erlangen",
@@ -1505,8 +504,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_erlangen_address,
         )
-
-        # Draft project
         self.project_erlangen_draft = Project.objects.create(
             name="Erlangen Draft Project",
             description="Draft project in Erlangen",
@@ -1518,8 +515,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_erlangen_address,
         )
-
-        # Negative test projects - Nuremberg
         self.project_nuremberg_location = Project.objects.create(
             name="Nuremberg Location Project",
             description="Project in Nuremberg location",
@@ -1531,7 +526,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_nuremberg,
         )
-
         self.project_nuremberg_address = Project.objects.create(
             name="Nuremberg Address Project",
             description="Project with address in Nuremberg",
@@ -1543,8 +537,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_nuremberg_address,
         )
-
-        # Negative test projects - Paris
         self.project_paris_location = Project.objects.create(
             name="Paris Location Project",
             description="Project in Paris location",
@@ -1556,7 +548,6 @@ class ProjectLocationHubFilterTest(TestCase):
             language=self.language,
             loc=self.location_paris,
         )
-
         self.project_paris_address = Project.objects.create(
             name="Paris Address Project",
             description="Project with address in Paris",
@@ -1571,10 +562,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_filter_projects_by_multi_location_hub(self):
-        """
-        Test filtering projects by the Erlangen hub with 3 locations.
-        Should return all 6 active, non-draft projects within Erlangen, Bubenreuth, and Spardorf.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1597,10 +584,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_aggregated_geometry_with_multiple_hub_locations(self):
-        """
-        Test the aggregated geometry functionality (Union of geometries).
-        Verifies that projects within any of the 3 hub locations are returned.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1610,17 +593,12 @@ class ProjectLocationHubFilterTest(TestCase):
         self.assertIn("erlangen-location-project", project_slugs)
         self.assertIn("bubenreuth-location-project", project_slugs)
         self.assertIn("spardorf-location-project", project_slugs)
-
         self.assertIn("erlangen-address-project", project_slugs)
         self.assertIn("bubenreuth-address-project", project_slugs)
         self.assertIn("spardorf-address-project", project_slugs)
 
     @tag("location_hub", "projects")
     def test_filter_projects_by_hub_filters_by_country(self):
-        """
-        Test that location hub filtering correctly filters by country first.
-        Paris projects should not appear even though we have 12 total projects.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1636,10 +614,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_filter_excludes_projects_outside_hub_geometry(self):
-        """
-        Test that projects outside the hub's geometry are excluded.
-        Nuremberg projects should not be included even though they're in Germany.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1648,15 +622,10 @@ class ProjectLocationHubFilterTest(TestCase):
 
         self.assertNotIn("nuremberg-location-project", project_slugs)
         self.assertNotIn("nuremberg-address-project", project_slugs)
-
         self.assertEqual(len(results), 6)
 
     @tag("location_hub", "projects")
     def test_filter_projects_by_nonexistent_hub(self):
-        """
-        Test filtering by a hub that doesn't exist.
-        Should return no results.
-        """
         response = self.client.get(self.url, {"hub": "nonexistent-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1665,10 +634,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_filter_projects_without_hub_parameter(self):
-        """
-        Test that without hub parameter, all active, non-draft projects are returned.
-        Should return 10 projects (12 total - 1 inactive - 1 draft).
-        """
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1682,11 +647,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_address_locations_within_hub_geometry(self):
-        """
-        Test that address-based locations (null multi_polygon) are correctly matched.
-        All 3 address projects should be included as their centre_points are within
-        the hub's aggregated geometry.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1707,10 +667,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_projects_annotated_with_distance(self):
-        """
-        Test that projects within a hub are annotated with distance.
-        All 6 projects should be returned with distance annotation.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1731,30 +687,17 @@ class ProjectLocationHubFilterTest(TestCase):
 
     @tag("location_hub", "projects")
     def test_distinct_results_when_multiple_filters_match(self):
-        """
-        Test that results are distinct even when a project matches multiple filter criteria.
-        Tests the .distinct() call in the code.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         results = response.json().get("results", [])
 
         project_ids = [p["id"] for p in results]
-        self.assertEqual(
-            len(project_ids),
-            len(set(project_ids)),
-            "Results should be distinct (no duplicates)",
-        )
-
+        self.assertEqual(len(project_ids), len(set(project_ids)))
         self.assertEqual(len(project_ids), 6)
 
     @tag("location_hub", "projects")
     def test_location_hub_filter_with_draft_projects(self):
-        """
-        Test that draft projects are not included in location hub filtering.
-        We have a draft project in Erlangen that should be excluded.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1762,15 +705,10 @@ class ProjectLocationHubFilterTest(TestCase):
 
         project_slugs = [p["url_slug"] for p in results]
         self.assertNotIn("erlangen-draft-project", project_slugs)
-
         self.assertEqual(len(results), 6)
 
     @tag("location_hub", "projects")
     def test_location_hub_filter_with_inactive_projects(self):
-        """
-        Test that inactive projects are not included in location hub filtering.
-        We have an inactive project in Erlangen that should be excluded.
-        """
         response = self.client.get(self.url, {"hub": "erlangen-hub"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -1778,7 +716,6 @@ class ProjectLocationHubFilterTest(TestCase):
 
         project_slugs = [p["url_slug"] for p in results]
         self.assertNotIn("erlangen-inactive-project", project_slugs)
-
         self.assertEqual(len(results), 6)
